+++ Obstacles/Sensor.cs
@@ -1,5 +1,5 @@
 namespace Obstacles;
 
 class Sensor : Obstacle{
-    public Sensor(byte x, byte y) : base(x, y) {}
+    public Sensor(int x, int y) : base(x, y) {}
 }
